+++ android/src/main/java/dev/t1r/themebuilder/android/MainActivity.kt
@@ -11,6 +11,8 @@ import androidx.lifecycle.lifecycleScope
 import com.arkivanov.decompose.defaultComponentContext
 import com.arkivanov.mvikotlin.logging.store.LoggingStoreFactory
 import com.arkivanov.mvikotlin.timetravel.store.TimeTravelStoreFactory
+import dev.t1r.themebuilder.component.materialcolorspalette.integration.MaterialColorsPaletteComponentImpl
+import dev.t1r.themebuilder.component.root.integration.RootComponentImpl
 import dev.t1r.themebuilder.data.ThemeBuilderDb
 import dev.t1r.themebuilder.data.colors.material.MaterialColorsDataSource
 import dev.t1r.themebuilder.data.colors.material.MaterialColorsRepositoryImpl
@@ -18,8 +20,6 @@ import dev.t1r.themebuilder.data.colors.theme.ThemeColorsDataSource
 import dev.t1r.themebuilder.data.colors.theme.ThemeColorsRepositoryImpl
 import dev.t1r.themebuilder.data.db.DriverFactory
 import dev.t1r.themebuilder.data.kvs.SettingsFactory
-import dev.t1r.themebuilder.feature.materialcolorspalette.integration.MaterialColorsPaletteComponentImpl
-import dev.t1r.themebuilder.feature.root.integration.RootComponentImpl
 import dev.t1r.themebuilder.repository.colors.theme.ThemeColorsRepository
 import dev.t1r.themebuilder.ui.compose.RootContent
 import kotlinx.coroutines.flow.collectLatest
