+++ common/component/baseline-colors/src/androidMain/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="dev.t1r.themebuilder.feature.baselinecolor"/>
+<manifest package="dev.t1r.themebuilder.component.baselinecolor"/>
