+++ common/component/baseline-colors/src/commonMain/kotlin/dev/t1r/themebuilder/component/baselinecolor/BaselineColorsComponent.kt
@@ -1,0 +1,18 @@
+package dev.t1r.themebuilder.component.baselinecolor
+
+import dev.t1r.themebuilder.entity.colors.ThemeColors
+import dev.t1r.themebuilder.entity.navigation.DrawerNavigationModel
+import kotlinx.coroutines.flow.Flow
+
+interface BaselineColorsComponent {
+    val models: Flow<Model>
+    val navigationModel: DrawerNavigationModel
+
+    data class Model(
+        val colors: ThemeColors = ThemeColors(),
+    )
+
+    data class Params(
+        val navigationModel: DrawerNavigationModel,
+    )
+}
