+++ common/component/baseline-colors/src/commonMain/kotlin/dev/t1r/themebuilder/component/baselinecolor/integration/BaselineColorsComponentImpl.kt
@@ -1,0 +1,26 @@
+package dev.t1r.themebuilder.component.baselinecolor.integration
+
+import com.arkivanov.decompose.ComponentContext
+import com.arkivanov.mvikotlin.core.store.StoreFactory
+import com.arkivanov.mvikotlin.extensions.coroutines.states
+import dev.t1r.themebuilder.component.baselinecolor.BaselineColorsComponent
+import dev.t1r.themebuilder.component.baselinecolor.BaselineColorsComponent.Model
+import dev.t1r.themebuilder.component.baselinecolor.store.BaselineColorsStoreProvider
+import dev.t1r.themebuilder.repository.colors.theme.ThemeColorsRepository
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
+
+class BaselineColorsComponentImpl constructor(
+    componentContext: ComponentContext,
+    storeFactory: StoreFactory,
+    params: BaselineColorsComponent.Params,
+    themeColorsRepository: ThemeColorsRepository,
+) : BaselineColorsComponent, ComponentContext by componentContext {
+    private val store = BaselineColorsStoreProvider(
+        storeFactory = storeFactory,
+        themeColorsRepository = themeColorsRepository,
+    ).provide()
+
+    override val models: Flow<Model> = store.states.map { stateToModel(it) }
+    override val navigationModel = params.navigationModel
+}
