--- common/component/baseline-colors/src/commonMain/kotlin/dev/t1r/themebuilder/feature/baselinecolor/BaselineColorsComponent.kt
@@ -1,18 +1,0 @@
-package dev.t1r.themebuilder.feature.baselinecolor
-
-import dev.t1r.themebuilder.entity.colors.ThemeColors
-import dev.t1r.themebuilder.entity.navigation.DrawerNavigationModel
-import kotlinx.coroutines.flow.Flow
-
-interface BaselineColorsComponent {
-    val models: Flow<Model>
-    val navigationModel: DrawerNavigationModel
-
-    data class Model(
-        val colors: ThemeColors = ThemeColors(),
-    )
-
-    data class Params(
-        val navigationModel: DrawerNavigationModel,
-    )
-}
