--- common/component/baseline-colors/src/commonMain/kotlin/dev/t1r/themebuilder/feature/baselinecolor/integration/UiMapper.kt
@@ -1,11 +1,0 @@
-package dev.t1r.themebuilder.feature.baselinecolor.integration
-
-import dev.t1r.themebuilder.feature.baselinecolor.BaselineColorsComponent
-import dev.t1r.themebuilder.feature.baselinecolor.store.BaselineColorsStore
-
-internal val stateToModel: BaselineColorsStore.State.() -> BaselineColorsComponent.Model =
-    {
-        BaselineColorsComponent.Model(
-            colors = model,
-        )
-    }
