+++ common/component/export/src/commonMain/kotlin/dev/t1r/themebuilder/component/export/ExportComponent.kt
@@ -1,0 +1,18 @@
+package dev.t1r.themebuilder.component.export
+
+import dev.t1r.themebuilder.entity.navigation.DrawerNavigationModel
+import kotlinx.coroutines.flow.Flow
+
+interface ExportComponent {
+    val models: Flow<Model>
+    val navigationModel: DrawerNavigationModel
+
+    data class Model(
+        val composeThemeExportString: String = "",
+        val androidXmlExportString: String = "",
+    )
+
+    data class Params(
+        val navigationModel: DrawerNavigationModel,
+    )
+}
