+++ common/component/export/src/commonMain/kotlin/dev/t1r/themebuilder/component/export/integration/UiMapper.kt
@@ -1,0 +1,12 @@
+package dev.t1r.themebuilder.component.export.integration
+
+import dev.t1r.themebuilder.component.export.ExportComponent
+import dev.t1r.themebuilder.component.export.store.ExportStore
+
+internal val stateToModel: ExportStore.State.() -> ExportComponent.Model =
+    {
+        ExportComponent.Model(
+            composeThemeExportString = composeThemeExportString,
+            androidXmlExportString = androidXmlExportString,
+        )
+    }
