+++ common/component/export/src/commonMain/kotlin/dev/t1r/themebuilder/component/export/store/ExportStoreProvider.kt
@@ -1,0 +1,122 @@
+package dev.t1r.themebuilder.component.export.store
+
+import com.arkivanov.mvikotlin.core.store.Reducer
+import com.arkivanov.mvikotlin.core.store.Store
+import com.arkivanov.mvikotlin.core.store.StoreFactory
+import com.arkivanov.mvikotlin.extensions.coroutines.CoroutineBootstrapper
+import com.arkivanov.mvikotlin.extensions.coroutines.CoroutineExecutor
+import dev.t1r.themebuilder.component.export.store.ExportStore.*
+import dev.t1r.themebuilder.entity.colors.ThemeColors
+import dev.t1r.themebuilder.repository.colors.theme.ThemeColorsRepository
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
+import kotlinx.coroutines.launch
+
+internal class ExportStoreProvider constructor(
+    private val storeFactory: StoreFactory,
+    private val themeColorsRepository: ThemeColorsRepository,
+) {
+
+    fun provide(): ExportStore = object : ExportStore, Store<Intent, State, Label> by storeFactory.create(
+        name = "ExportStore",
+        initialState = State(),
+        bootstrapper = BootstrapperImpl(themeColorsRepository),
+        executorFactory = ::ExecutorImpl,
+        reducer = ReducerImpl,
+    ) {}
+
+    private sealed class Message {
+        data class UpdateComposeThemeExportString(val exportString: String) : Message()
+        data class UpdateAndroidXmlExportString(val exportString: String) : Message()
+    }
+
+    private inner class ExecutorImpl : CoroutineExecutor<Intent, Action, State, Message, Label>() {
+        override fun executeAction(
+            action: Action,
+            getState: () -> State,
+        ): Unit = when (action) {
+            is Action.UpdateColors -> handleUpdateColors(action.model)
+        }
+
+        private fun handleUpdateColors(model: ThemeColors) {
+            updateComposeThemeExport(model)
+            updateAndroidXmlThemeExport(model)
+        }
+
+        private fun updateComposeThemeExport(model: ThemeColors) {
+            val exportString = """
+MaterialTheme(
+  colors = Colors(
+    primary = ${mapToComposeColorString(model.primary)},
+    primaryVariant = ${mapToComposeColorString(model.primaryVariant)},
+    secondary = ${mapToComposeColorString(model.secondary)},
+    secondaryVariant = ${mapToComposeColorString(model.secondaryVariant)},
+    background = ${mapToComposeColorString(model.background)},
+    surface = ${mapToComposeColorString(model.surface)},
+    error = ${mapToComposeColorString(model.error)},
+    onPrimary = ${mapToComposeColorString(model.onPrimary)},
+    onSecondary = ${mapToComposeColorString(model.onSecondary)},
+    onBackground = ${mapToComposeColorString(model.onBackground)} ,
+    onSurface = ${mapToComposeColorString(model.onSurface)},
+    onError = ${mapToComposeColorString(model.onError)},
+    isLight = ${model.isLight},
+  ),
+  content = {},
+)
+""".trimIndent()
+            dispatch(Message.UpdateComposeThemeExportString(exportString))
+        }
+
+        private fun updateAndroidXmlThemeExport(model: ThemeColors) {
+            val exportString = """
+<!-- Use com.google.android.material:material:1.1.0 and higher -->
+<style name="Theme.ApplicationName" parent="Theme.MaterialComponents.NoActionBar">
+    <item name="colorPrimary">${mapToAndroidXmlColorString(model.primary)}</item>
+    <item name="colorPrimaryVariant">${mapToAndroidXmlColorString(model.primaryVariant)}</item>
+    <item name="colorSecondary">${mapToAndroidXmlColorString(model.secondary)}</item>
+    <item name="colorSecondaryVariant">${mapToAndroidXmlColorString(model.secondaryVariant)}</item>
+    <item name="android:colorBackground">${mapToAndroidXmlColorString(model.background)}</item>
+    <item name="colorSurface">${mapToAndroidXmlColorString(model.surface)}</item>
+    <item name="colorError">${mapToAndroidXmlColorString(model.error)}</item>
+    <item name="colorOnPrimary">${mapToAndroidXmlColorString(model.onPrimary)}</item>
+    <item name="colorOnSecondary">${mapToAndroidXmlColorString(model.onSecondary)}</item>
+    <item name="colorOnBackground">${mapToAndroidXmlColorString(model.onBackground)} </item>
+    <item name="colorOnSurface">${mapToAndroidXmlColorString(model.onSurface)}</item>
+    <item name="colorOnError">${mapToAndroidXmlColorString(model.onError)}</item>
+    <item name="android:statusBarColor">?attr/colorPrimaryVariant</item>
+</style>
+""".trimIndent()
+            dispatch(Message.UpdateAndroidXmlExportString(exportString))
+        }
+
+        private fun mapToComposeColorString(from: Long): String =
+            "Color(0x${from.toString(16)})"
+
+        private fun mapToAndroidXmlColorString(from: Long): String =
+            "#${from.toString(16)}"
+    }
+
+    private object ReducerImpl : Reducer<State, Message> {
+        override fun State.reduce(msg: Message): State = when (msg) {
+            is Message.UpdateComposeThemeExportString -> copy(
+                composeThemeExportString = msg.exportString,
+            )
+
+            is Message.UpdateAndroidXmlExportString -> copy(
+                androidXmlExportString = msg.exportString,
+            )
+        }
+    }
+
+    private class BootstrapperImpl(
+        private val themeColorsRepository: ThemeColorsRepository,
+    ) : CoroutineBootstrapper<Action>() {
+        override fun invoke() {
+            scope.launch {
+                themeColorsRepository.themeColorsState()
+                    .onEach { dispatch(Action.UpdateColors(it)) }
+                    .launchIn(this)
+            }
+        }
+    }
+}
