--- common/component/export/src/commonMain/kotlin/dev/t1r/themebuilder/feature/export/integration/UiMapper.kt
@@ -1,12 +1,0 @@
-package dev.t1r.themebuilder.feature.export.integration
-
-import dev.t1r.themebuilder.feature.export.ExportComponent
-import dev.t1r.themebuilder.feature.export.store.ExportStore
-
-internal val stateToModel: ExportStore.State.() -> ExportComponent.Model =
-    {
-        ExportComponent.Model(
-            composeThemeExportString = composeThemeExportString,
-            androidXmlExportString = androidXmlExportString,
-        )
-    }
