+++ common/component/material-colors-palette/src/androidMain/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="dev.t1r.themebuilder.feature.materialcolorspalette"/>
+<manifest package="dev.t1r.themebuilder.component.materialcolorspalette"/>
