--- common/component/material-colors-palette/src/commonMain/kotlin/dev/t1r/themebuilder/feature/materialcolorspalette/integration/EventMapper.kt
@@ -1,10 +1,0 @@
-package dev.t1r.themebuilder.feature.materialcolorspalette.integration
-
-import  dev.t1r.themebuilder.feature.materialcolorspalette.MaterialColorsPaletteComponent.Event
-import dev.t1r.themebuilder.feature.materialcolorspalette.store.MaterialColorsPaletteStore
-
-internal val labelToEvent: MaterialColorsPaletteStore.Label.() -> Event = {
-    when (this) {
-        is MaterialColorsPaletteStore.Label.Error -> Event.Error
-    }
-}
