+++ common/component/root/src/androidMain/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="dev.t1r.themebuilder.feature.root"/>
+<manifest package="dev.t1r.themebuilder.component.root"/>
