+++ common/component/root/src/commonMain/kotlin/dev/t1r/themebuilder/component/root/integration/UiMapper.kt
@@ -1,0 +1,11 @@
+package dev.t1r.themebuilder.component.root.integration
+
+import dev.t1r.themebuilder.component.root.RootComponent
+import dev.t1r.themebuilder.component.root.store.RootStore
+
+internal val stateToModel: RootStore.State.() -> RootComponent.Model =
+    {
+        RootComponent.Model(
+            colors = model,
+        )
+    }
