+++ common/component/root/src/commonMain/kotlin/dev/t1r/themebuilder/component/root/store/RootStore.kt
@@ -1,0 +1,18 @@
+package dev.t1r.themebuilder.component.root.store
+
+import com.arkivanov.mvikotlin.core.store.Store
+import dev.t1r.themebuilder.component.root.store.RootStore.*
+import dev.t1r.themebuilder.entity.colors.ThemeColors
+
+interface RootStore : Store<Intent, State, Label> {
+    data class State(
+        val model: ThemeColors = ThemeColors(),
+    )
+
+    sealed class Action {
+        data class UpdateColors(val model: ThemeColors) : Action()
+    }
+
+    sealed class Intent
+    sealed class Label
+}
