--- common/component/root/src/commonMain/kotlin/dev/t1r/themebuilder/feature/root/integration/UiMapper.kt
@@ -1,11 +1,0 @@
-package dev.t1r.themebuilder.feature.root.integration
-
-import dev.t1r.themebuilder.feature.root.RootComponent
-import dev.t1r.themebuilder.feature.root.store.RootStore
-
-internal val stateToModel: RootStore.State.() -> RootComponent.Model =
-    {
-        RootComponent.Model(
-            colors = model,
-        )
-    }
