+++ common/ui-compose/src/commonMain/kotlin/dev/t1r/themebuilder/ui/compose/BaselineColorsUi.kt
@@ -16,9 +16,9 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import dev.t1r.themebuilder.feature.baselinecolor.BaselineColorsComponent
-import dev.t1r.themebuilder.feature.baselinecolor.BaselineColorsComponent.Model
-import dev.t1r.themebuilder.feature.materialcolorspalette.MaterialColorsPaletteComponent
+import dev.t1r.themebuilder.component.baselinecolor.BaselineColorsComponent
+import dev.t1r.themebuilder.component.baselinecolor.BaselineColorsComponent.Model
+import dev.t1r.themebuilder.component.materialcolorspalette.MaterialColorsPaletteComponent
 import dev.t1r.themebuilder.ui.compose.common.ColorsScreenContainerWidget
 
 @Composable
