+++ common/ui-compose/src/commonMain/kotlin/dev/t1r/themebuilder/ui/compose/common/ColorsScreenContainer.kt
@@ -15,8 +15,8 @@ import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.unit.dp
+import dev.t1r.themebuilder.component.materialcolorspalette.MaterialColorsPaletteComponent
 import dev.t1r.themebuilder.entity.navigation.DrawerNavigationModel
-import dev.t1r.themebuilder.feature.materialcolorspalette.MaterialColorsPaletteComponent
 import dev.t1r.themebuilder.ui.compose.appmenu.AppMenuWidget
 import dev.t1r.themebuilder.ui.compose.materialcolorspalette.MaterialColorsPaletteContent
 import kotlinx.coroutines.launch
